refactor: add utils
